Use long indexes.

+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/numeric/arithmetic/MaxOp.java
@@ -13,18 +13,18 @@ public class MaxOp<T extends Number, I extends Storage<? super T>>
 
   @Override
   public double doDouble(
-      double a, double b, int ix, MapOperationProblemAggregator problemAggregator) {
+      double a, double b, long ix, MapOperationProblemAggregator problemAggregator) {
     return Math.max(a, b);
   }
 
   @Override
-  public Long doLong(long a, long b, int ix, MapOperationProblemAggregator problemAggregator) {
+  public Long doLong(long a, long b, long ix, MapOperationProblemAggregator problemAggregator) {
     return Math.max(a, b);
   }
 
   @Override
   public BigInteger doBigInteger(
-      BigInteger a, BigInteger b, int ix, MapOperationProblemAggregator problemAggregator) {
+      BigInteger a, BigInteger b, long ix, MapOperationProblemAggregator problemAggregator) {
     return a.max(b);
   }
